Removed unwanted joins, make direct table call.
As tables grow bigger, joins become very slow. Calling the table directly improves performance, especially with the new indexes.

Added a scope association (hours_acceptance_breaks_enabled) in HoursAcceptancePeriod to be able to use eager loading for that scope. Because eager loading scopes does not work, but works with associations.